+++ packages/algob/src/errors/errors.ts
@@ -35,7 +35,7 @@ export class BuilderError extends Error {
       messageArguments
     );
 
-    super(prefix + formattedMessage);
+    super(String(prefix) + String(formattedMessage));
     this.errorDescriptor = errorDescriptor;
     this.number = errorDescriptor.number;